+++ client/src/dialogs/account-dialog.js
@@ -8,10 +8,40 @@ import { I18N } from "aurelia-i18n";
 import QConfig from "resources/QConfig.js";
 import User from "resources/User.js";
 import Auth from "resources/Auth.js";
+import ToolsInfo from "resources/ToolsInfo.js";
 
-@inject(DialogController, Container, Router, I18N, User, Auth, QConfig)
+@inject(
+  DialogController,
+  Container,
+  Router,
+  I18N,
+  User,
+  Auth,
+  QConfig,
+  ToolsInfo
+)
 export class AccountDialog {
-  constructor(controller, diContainer, router, i18n, user, auth, qConfig) {
+  toolSelectionTypes = [
+    {
+      value: "byUsage",
+      label_i18n_key: "userConfig.toolSelectionType.byUsage"
+    },
+    {
+      value: "byConfig",
+      label_i18n_key: "userConfig.toolSelectionType.byConfig"
+    }
+  ];
+
+  constructor(
+    controller,
+    diContainer,
+    router,
+    i18n,
+    user,
+    auth,
+    qConfig,
+    toolsInfo
+  ) {
     this.controller = controller;
     this.diContainer = diContainer;
     this.router = router;
@@ -19,6 +49,7 @@ export class AccountDialog {
     this.user = user;
     this.auth = auth;
     this.qConfig = qConfig;
+    this.toolsInfo = toolsInfo;
 
     this.loadData();
   }
@@ -34,6 +65,40 @@ export class AccountDialog {
     await this.user.loaded;
     if (!this.user.isLoggedIn) {
       this.router.navigateToRoute("login");
+    }
+
+    await this.loadToolSelectionConfig();
+  }
+
+  async loadToolSelectionConfig() {
+    this.tools = await this.toolsInfo.getAvailableTools();
+
+    this.toolSelectionConfig = this.user.getUserConfig("toolSelection", {
+      type: "byUsage",
+      tools: {}
+    });
+    // check if the user has any tools selected for the initial tool selection ...
+    let hasAnyToolSelectedForInitialToolSelection = false;
+    for (const toolName in this.toolSelectionConfig.tools) {
+      if (
+        this.toolSelectionConfig.tools[toolName].inInitialToolSelection === true
+      ) {
+        hasAnyToolSelectedForInitialToolSelection = true;
+        break;
+      }
+    }
+    // .. if not we set the default to the first 4 most used ones
+    if (!hasAnyToolSelectedForInitialToolSelection) {
+      for (let i = 0; i < 4; i++) {
+        if (
+          !this.toolSelectionConfig.tools.hasOwnProperty(this.tools[i].name)
+        ) {
+          this.toolSelectionConfig.tools[this.tools[i].name] = {};
+        }
+        this.toolSelectionConfig.tools[
+          this.tools[i].name
+        ].inInitialToolSelection = true;
+      }
     }
   }