+++ client/src/elements/molecules/index.js
@@ -1,9 +1,9 @@
 export function configure(frameworkConfiguration) {
   frameworkConfiguration
-    .globalResources("./confirm-button")
-    .globalResources("./delete-button.html")
     .globalResources("./embed-code")
     .globalResources("./item-list-entry")
+    .globalResources("./dropdown-menu")
     .globalResources("./language-switcher")
-    .globalResources("./radio-button-group");
+    .globalResources("./radio-button-group")
+    .globalResources("./notification");
 }