+++ client/src/elements/molecules/notification.js
@@ -1,0 +1,17 @@
+import { bindable } from "aurelia-framework";
+
+export class Notification {
+  @bindable notification;
+
+  hasIcon = false;
+  iconName = "";
+
+  notificationChanged() {
+    if (this.notification && this.notification.priority) {
+      this.hasIcon = ["medium", "high"].includes(
+        this.notification.priority.type
+      );
+      this.iconName = `notification-${this.notification.priority.type}`;
+    }
+  }
+}