+++ client/src/elements/organisms/tool-status-bar.js
@@ -1,27 +1,41 @@
 import { bindable, inject } from "aurelia-framework";
+import { BindingSignaler } from "aurelia-templating-resources";
 import { Notification } from "aurelia-notification";
 import { DialogService } from "aurelia-dialog";
 import { ItemDialog } from "dialogs/item-dialog.js";
 import QConfig from "resources/QConfig.js";
 
-@inject(QConfig, Notification, DialogService)
+@inject(QConfig, Notification, DialogService, BindingSignaler)
 export class ToolStatusBar {
   @bindable item;
   @bindable saveAction;
   @bindable activateAction;
+  @bindable lastSavedDate;
 
   message;
 
-  constructor(qConfig, notification, dialogService) {
+  constructor(qConfig, notification, dialogService, bindingSignaler) {
     this.qConfig = qConfig;
     this.notification = notification;
     this.dialogService = dialogService;
+    this.bindingSignaler = bindingSignaler;
   }
 
   created() {
     return this.qConfig.get("uiBehavior").then(uiBehaviorConfig => {
       this.uiBehaviorConfig = uiBehaviorConfig;
     });
+  }
+
+  attached() {
+    this.updateTimeAgoIntervalId = setInterval(
+      () => this.bindingSignaler.signal("update-timeago"),
+      1000
+    );
+  }
+
+  detached() {
+    clearInterval(this.updateTimeAgoIntervalId);
   }
 
   onActivateClick() {