+++ client/src/resources/User.js
@@ -1,9 +1,13 @@
-import { inject } from "aurelia-framework";
+import { inject, LogManager } from "aurelia-framework";
 import { HttpClient } from "aurelia-fetch-client";
 import qEnv from "resources/qEnv.js";
 
+const log = LogManager.getLogger("Q");
+
 @inject(HttpClient)
 export default class User {
+  changeCallbacks = [];
+
   constructor(httpClient) {
     this.httpClient = httpClient;
     this.isLoggedIn = false;
@@ -43,11 +47,20 @@ export default class User {
     this.save();
   }
 
-  getUserConfig(key) {
-    if (!this.data || !this.data.hasOwnProperty("config")) {
-      return null;
+  getUserConfig(key, defaultValue = undefined) {
+    if (
+      (!this.data ||
+        !this.data.hasOwnProperty("config") ||
+        !this.data.config.hasOwnProperty(key) ||
+        this.data.config[key] === undefined) &&
+      defaultValue !== undefined
+    ) {
+      this.data.config[key] = defaultValue;
     }
-    return this.data.config[key];
+    if (this.data && this.data.config && this.data.config[key]) {
+      return this.data.config[key];
+    }
+    return undefined;
   }
 
   get roles() {
@@ -76,9 +89,27 @@ export default class User {
       if (!response.ok) {
         throw response;
       }
+      try {
+        for (let cb of this.changeCallbacks) {
+          cb();
+        }
+      } catch (e) {
+        log.error("failed to call user change callback", e);
+      }
       return true;
     } catch (e) {
       return false;
+    }
+  }
+
+  registerChangeCallback(cb) {
+    this.changeCallbacks.push(cb);
+  }
+
+  unregisterChangeCallback(cb) {
+    const index = this.changeCallbacks.indexOf(cb);
+    if (index > -1) {
+      this.changeCallbacks.splice(index, 1);
     }
   }
 }