+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+# 3.0.0 (27.06.2018)
+- feat: notificationChecks, Q Editor can show notifications in the schema-editor to help the user enter correct data. These checks are configured in the tools schema.json
+- feat: tool-selection can no be configured by the user and the order can be based on the users usage (needs Q-server >= 5)
+- feat: item-list on index page contains a menu per item to delete/edit/blueprint directly
+- refresh the overall visual appearance
+- improvement: item-preview widths are now configurable together with min-height per width
+- improvement: tool schema translation translates placeholder in Q:options as well
+- fix: schema-editor-table allows to copy 10000 rows now
+- breaking: some CSS custom property names changed
+- breaking: translations for tools and editor config are separated (reflected in Q-server@5)
+
 # 2.3.0 (29.05.2018)
 - feat: schema-editor-table minRows is now configurable via Q:options in schema.json
 - fix: schema-editor takes the whole width and expands rows to do so